Refactor: (상대방)과 (사용자)간의 숫자 비교 로직 클래스 분리
feat:
- Compare.js 파일을 새로 생성하여 숫자 비교 로직을 따로 분리해 구현하였습니다.
- 각 클래스의 책임을 명확히 구분해 로직을 효과적으로 관리할 수 있도록 하기 위해서 해당 리팩토링 과정을 진행하였습니다.
